+++ templates/ProtobuffEncoder.Template.Normal/Contracts/Team.cs
@@ -1,0 +1,11 @@
+using ProtobuffEncoder.Attributes;
+
+[ProtoContract]
+public class Team
+{
+    [ProtoField(1)] public string Name { get; set; } = "";
+    [ProtoField(2)] public List<string> Members { get; set; } = [];
+
+    [ProtoMap]
+    [ProtoField(3)] public Dictionary<string, int> Scores { get; set; } = new();
+}